@@ -1,18 +1,15 @@
 import { NG } from '@web-companions/gfc';
+import { litView } from '@web-companions/lit';
 import { svg } from 'lit-html';
-import { renderNode } from './utils/directives';
 
-export const iconNote = NG(function* () {
+export const iconNote = litView.node(function* () {
   while (true) {
-    yield renderNode(
-      svg/*html*/ `
+    yield svg/*html*/ `
         <svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-pencil" width="24" height="24" viewBox="0 0 24 24" stroke-width="1" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
           <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
           <path d="M4 20h4l10.5 -10.5a1.5 1.5 0 0 0 -4 -4l-10.5 10.5v4"></path>
           <line x1="13.5" y1="6.5" x2="17.5" y2="10.5"></line>
         </svg>
-      `,
-      this
-    );
+      `;
   }
 });
