@@ -1,3 +1,4 @@
-export { EG } from './element';
-export { NG } from './node';
-export { p } from './utils';
+export { EG } from './EG';
+export { NG } from './NG';
+export { View } from './View';
+export { p } from './utils/p';
